fix: match the monkey-patched method signature with upstream openai
When instructor monkey-patches the create and acreate methods, it changes
the type of the function from 'classmethod' to 'function'. This works fine
in most cases, but when other tools attempt to, themselves, wrap the OpenAI
class methods, this produces an error because those tools are expecting
classmethods; not functions.

The example I encountered is using the python "wrapt" library/module. When
wrapping the monkey-patched code, it produces the following error

  TypeError: missing 1 required positional argument

This patch changes the method signature to match the one that OpenAI themselves
use.

I've included tests to ensure that the existing instructor behavior remains
unchanged. Additionally, I tested with the tests that are marked with pytest
unimplemented marks to ensure that the continued class instances are correct.