Anthropic client. Prevent empty generator when streaming
This commit fixes #727.

When using the Anthropic client in streaming (stream=true)
mode, an empty generator was returned, even though
the LLM model response was non-empty. This was
caused by completion text being looked for in
the wrong place on the Anthopic SDK's completion chunk
objection. (I'm guessing something changed in an update.)
In the method `extract_json()` in dsl/iterable.py,
this is where completion text was looked for previously:

```
chunk.model_extra.get("delta", "").get("partial_json", "")
```

The attribute `model_extra`, however, is an empty
dict. The first `.get()` therefore gives a string,
causing the second `.get()` to raise an AttributeError.
That exception was silently swallowed, since the
whole block is wrapped in a try-except, and
all chunks were ignored.

In the Antropic SDK, `delta` is now an attribute
and no dicts are involved.

Updating the line above as follows fixes the issue:

```
chunk.delta.partial_json
```

I've applied this fix in both iterable.py
and partial.py. I've tested both with `Iterable[]`
and `Partial[]` and with with modes ANTHROPIC_TOOLS
and ANTHROPIC_JSON. I've also tested `create_iterable()`
and `create_partial()`. While testing the last part
I removed the guard that prevented Anthropic being
used in this mode. Everything seems to work fine,
so I've dropped it.